requirements.txt
Django==2.1.2
djangorestframework==3.9.0
dj-database-url==0.5.0
psycopg2==2.7.5